partial solution of mid
